Add scaffold for chat rooms resource
In this commit I want to add a feature of CRUD chat
rooms for our pheonix users. This way, the admin can
create topic rooms for users talk.

For example: Admin can create a "Geek" room for people
that like games, hq and etc. And also create a room
"Programming" for people that like to discuss programming
techniques.

A scaffold is useful to quick generating something
that works. Phoenix is so awesome that it generates
also with testing code \o/. For someone that is learning
helps a lot.

Even using a scaffold, is necessary manual intervention
for example to add constraints to your database(it is very important!).

Our rooms must have unique names in the database, that's why I've changed
the migration file adding the unique index. And also we don't need to
store large names for our chat rooms.