Add basic http authentication for admin area.
We don't want to everyone have access to manipulate
our chart rooms. With this commit we have add a lib
that provides the basic http authentication. And
plugged it in route scope called "admin".

Then if you want manage the rooms you need to access:

/admin/rooms

You'll be prompted to login, just provide the
admin username and password configured on
config.exs file.