Refine the simple web chat app.
Well, sometimes we want to improve an
application, when the times comes we need
to refactor and create abstracntions to
proper envolver without too much pain.

The next step is permit user select
a room, then I created a ChatRoom class
in JS side to represent it.

Also add some UX improvements, like
scroll to message bottom when we have
too many messages, also added a scroll
when we have too much users.